+++ lib/screens/authentication/authentication.dart
@@ -1,7 +1,5 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:ecochat_app/services/markers_api.dart';
-
 import '../homepage/homepage.dart';
 
 class AuthenticationView extends StatefulWidget {
@@ -16,6 +14,7 @@ class _AuthenticationViewState extends State<AuthenticationView> {
   final TextEditingController _emailField = TextEditingController();
   final TextEditingController _passwordField = TextEditingController();
   // late MarkersAPI markersAPI;
+  final bool shouldNavigate = false;
 
   @override
   Widget build(BuildContext context) {
