+++ pubspec.yaml
@@ -32,11 +32,6 @@ dependencies:
     sdk: flutter
   meta: ^1.7.0
   signalr_netcore: 1.2.3
-
-
-
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
 
 dev_dependencies:
